Move availablePaymentGateways to checkout query
