+++ saleor/graphql/shop/types.py
@@ -17,7 +17,6 @@ from ..core.enums import WeightUnitsEnum
 from ..core.types.common import (
     CountryDisplay, LanguageDisplay, PermissionDisplay)
 from ..menu.types import Menu
-from ..payment.enums import PaymentGatewayEnum
 from ..product.types import Collection
 from ..utils import format_permissions_for_display
 from .enums import AuthorizationKeyType
@@ -77,9 +76,6 @@ class Shop(graphene.ObjectType):
         required=True)
     default_currency = graphene.String(
         description='Default shop\'s currency.', required=True)
-    available_payment_gateways = graphene.List(
-        PaymentGatewayEnum, description='List of available payment gateways.',
-        required=True)
     default_country = graphene.Field(
         CountryDisplay, description='Default shop\'s country')
     description = graphene.String(description='Shop\'s description.')
@@ -147,9 +143,6 @@ class Shop(graphene.ObjectType):
 
     def resolve_default_currency(self, info):
         return settings.DEFAULT_CURRENCY
-
-    def resolve_available_payment_gateways(self, info):
-        return settings.CHECKOUT_PAYMENT_GATEWAYS.keys()
 
     def resolve_description(self, info):
         return info.context.site.settings.description
